GameState compression: make `Piece` (and therefore `GameState.CapturedPiece`) a sbyte enum

+++ src/Lynx/Model/GameState.cs
@@ -1,7 +1,7 @@
 ﻿namespace Lynx.Model;
 public readonly struct GameState
 {
-    public readonly int CapturedPiece;
+    public readonly sbyte CapturedPiece;
 
     public readonly byte Castle;
 
@@ -11,7 +11,7 @@ public readonly struct GameState
 
     // TODO: save full Zobrist key?
 
-    public GameState(int capturedPiece, byte castle, BoardSquare enpassant, long zobristKey)
+    public GameState(sbyte capturedPiece, byte castle, BoardSquare enpassant, long zobristKey)
     {
         CapturedPiece = capturedPiece;
         Castle = castle;
